4300 - Long in int

+++ Logistics.API/Logisitcs.DAL/Models/Status.cs
@@ -4,7 +4,7 @@ namespace Logisitcs.DAL.Models;
 
 public partial class Status
 {
-    public long StatusId { get; set; }
+    public int StatusId { get; set; }
 
     public string Name { get; set; }
